pyproject.toml
[project]
name = "ai-real-estate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "joblib>=1.5.3",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "scikit-learn==1.6.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
